Move library to root of repository
A popular library installation technique is to download the library via GitHub's Clone or download > Download ZIP and then use the Arduino IDE's Sketch > Include Library > Add .ZIP Library on the downloaded file. This requires the library to be in the root of the repository, not in a subfolder. If the library is not in the root of the repository this installation technique fails:

Specified folder/zip file does not contain a valid library

This is the standard repository structure used in all official Arduino libraries:
https://github.com/arduino-libraries

This change also allows the example sketches to be accessed via the File > Examples > MagzorI2C menu after installing the library.

This move is also required if you wanted to add your library to the Arduino Library Manager index, which provides an even easier installation option.